Fix mypy [index]
Before, self.curr_inst_idx only was of type int, but now is of type int and float, because it can get assigned np.inf.
From this, the index error arises, when mypy thinks we want to index a list with np.inf.
However, np.inf only might be assigned later.