Don't use mutable default argument.
If this isn't respected, then additional_configs can be built up
overtime when multiple optimizers are created in a single process.

This can occur when the use_default_config option is set True and/or
when additional_configs are provided.

When different ConfigSpaces are provided this can result in an error due
to mismatched defaults, for instance.

In other cases it can result in non-deterministic behavior due to the
random ordering of the configs that get added (e.g., with different
parallel pytest runs).

- [ ] Cleanup other sources of this error as well elsewhere in the code.
- [ ] Add pylint checks for this issue.